nhl-game-predictor-django-backend/pyproject.toml
[tool.poetry]
name = "nhl-game-predictor-backend"
version = "0.1.0"
description = ""
authors = ["Evan Boyd <evan.boyd@mail.mcgill.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.2"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
httpx = "^0.27.2"
datetime = "^5.5"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
djangorestframework = "^3.15.2"
lime = "^0.2.0.1"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
pytz = "^2024.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
